gcp_cloud_functions/copy_blobs_gcp_storage/requirements.txt
-i https://pypi.org/simple
google-cloud-storage==1.13.1
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-api-core==1.11.1
requests==2.22.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
google-auth==1.6.3
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
protobuf==3.8.0
pyasn1-modules==0.2.5
pyasn1==0.4.5
pytz==2019.1
rsa==4.0
six==1.12.0
urllib3==1.25.3
pandas==1.2.3
fsspec==0.8.7
gcsfs==0.7.2

